Some error catches, dynamic path. Some minor changes

+++ index.js
@@ -1,45 +1,45 @@
 require("dotenv").config();
 const Discord = require("discord.js");
-const client = new Discord.Client();
-
+const path = require("path");
 const fs = require("fs");
 const TikTokScraper = require("tiktok-scraper");
 const config = require("./config.json");
-const command = require("./command");
+const util = require("./util");
+const client = new Discord.Client();
+const confirmReaction = ["✅", "❌"];
+
+module.exports = { client };
+const addReactions = (message, reactions) => {
+  message.react(reactions[0]);
+  reactions.shift();
+  if (reactions.length > 0) {
+    setTimeout(() => addReactions(message, reactions), 750);
+  }
+};
 
 client.on("ready", () => {
+  var dir = __dirname + "/temp/";
+
+  if (!fs.existsSync(dir)) {
+    fs.mkdirSync(dir);
+  }
   console.log(`Logged in as ${client.user.tag}!`);
-  command(client, ["ping", "test"], (message) => {
-    message.channel.send("TEEE");
-  });
-  command(client, "servers", (message) => {
-    client.guilds.cache.forEach((guild) => {
-      console.log(guild);
-      message.channel.send(
-        `${guild.name} has a total of ${guild.memberCount} members`
-      );
-    });
-  });
-  command(client, ["cc", "clearchannnel", "delMsg"], (message) => {
-    if (message.member.hasPermission("ADMINISTRATOR")) {
-      message.channel.messages.fetch().then((results) => {
-        //message.channel.bulkDelete(results);
-        message.channel.send(
-          "Are you sure you want to delete everything in this channel ?"
-        );
-        //console.log(results);
-      });
+  const baseFile = "command-base.js";
+  const commandBase = require(`./commands/${baseFile}`);
+
+  const readCommands = (dir) => {
+    const files = fs.readdirSync(path.join(__dirname, dir));
+    for (const file of files) {
+      const stat = fs.lstatSync(path.join(__dirname, dir, file));
+      if (stat.isDirectory()) {
+        readCommands(path.join(dir, file));
+      } else if (file !== baseFile) {
+        const option = require(path.join(__dirname, dir, file));
+        commandBase(client, option);
+      }
     }
-  });
-  command(client, "status", (message) => {
-    const content = message.content.replace("!status ", "");
-    client.user.setPresence({
-      activity: {
-        name: content,
-        type: 0,
-      },
-    });
-  });
+  };
+  readCommands("commands");
 });
 
 let scrapperMsg;
@@ -59,14 +59,22 @@ client.on("message", async (msg) => {
     var expression = /(https?:\/\/(?:www\.|(?!www))[^\s\.]+\.[^\s]{2,}|www\.[^\s]+\.[^\s]{2,})/gi;
     let message = msg.content.match(expression);
     await startScraper(message);
-    msg.delete();
-    const buffer = fs.readFileSync(String(scrapperMsg[2]));
-    const attachment = new Discord.MessageAttachment(buffer, "YourVideo.mp4");
-    msg.reply(attachment);
+    if (getFilesizeInMB(String(scrapperMsg[2])) > 8) {
+      msg.reply(
+        "Sorry the Video you asked for is bigger than 8MB, I do not support compression yet!"
+      );
+      return;
+    } else {
+      msg.delete();
+      const buffer = fs.readFileSync(String(scrapperMsg[2]));
+      const attachment = new Discord.MessageAttachment(buffer, "YourVideo.mp4");
+      msg.reply(attachment);
+    }
   }
 });
 
 client.login(process.env.BOTTOKEN); // Add your DiscordBot Tokken here @@@@@
+
 const startScraper = async (link) => {
   try {
     const scraper = await TikTokScraper["video"](link, options);
